# docs/openapi.json
{
...
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "description": "Stable SCREAMING_SNAKE_CASE machine string. Once shipped, never reused.",
            "type": "string"
          },
          "details": {
            "description": "Optional structured detail. Omitted on auth failures per INV-A4.",
            "nullable": true
          },
          "message": {
            "description": "Human-readable, lowercase, no trailing period.",
            "type": "string"
          },
          "request_id": {
            "description": "UUIDv7 request correlator echoed in the `X-Request-Id` header.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "type": "object"
      },
...
      "OriginDto": {
        "properties": {
          "id": {
            "description": "`cli:<pid>` → \"<pid>\" as a string; `service:<session_id>` → session id.",
            "type": "string"
          },
          "kind": {
            "description": "`cli` | `service`.",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "id"
        ],
        "type": "object"
      },
...
      "SliceCountsDto": {
        "properties": {
          "blocked_retry": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "completed": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "escalated": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "in_progress": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "pending": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "refused": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "pending",
          "in_progress",
          "blocked_retry",
          "completed",
          "escalated",
          "refused"
        ],
        "type": "object"
      },
      "SliceDto": {
        "properties": {
          "attempts": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "author_agent": {
            "type": "string"
          },
          "bounded_context": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "layer": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "objective": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "target_loc": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "layer",
          "bounded_context",
          "author_agent",
          "attempts",
          "target_loc",
          "objective"
        ],
        "type": "object"
      },
      "SliceQueueDto": {
        "properties": {
          "generated_at": {
            "type": "string"
          },
          "pipeline_mode": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "slices": {
            "items": {
              "$ref": "#/components/schemas/SliceDto"
            },
            "type": "array"
          }
        },
        "required": [
          "project_id",
          "pipeline_mode",
          "generated_at",
          "slices"
        ],
        "type": "object"
      },
      "StateLogPageDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StateUpdateDto"
            },
            "type": "array"
          },
          "next_cursor": {
            "description": "Opaque cursor (DSD-625) — base64url of the next byte offset. Absent at tail.",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "StateUpdateDto": {
        "properties": {
          "at": {
            "type": "string"
          },
          "body": {
            "nullable": true
          },
          "event": {
            "type": "string"
          },
          "origin": {
            "allOf": [
              {
                "$ref": "#/components/schemas/OriginDto"
              }
            ],
            "nullable": true
          },
          "schema_version": {
            "format": "int32",
            "minimum": 0,
            "nullable": true,
            "type": "integer"
          },
          "slice_id": {
            "type": "string"
          }
        },
        "required": [
          "slice_id",
          "event",
          "at"
        ],
        "type": "object"
      },
      "StatusDto": {
        "properties": {
          "active_slice_id": {
            "description": "Slice currently held in `.mutagen/state/active-slice.json`, if any.",
            "nullable": true,
            "type": "string"
          },
          "active_stage": {
            "nullable": true,
            "type": "string"
          },
          "last_state_update_at": {
            "description": "Wall-clock of the last State Update record appended to the log (DSD-622).",
            "nullable": true,
            "type": "string"
          },
          "pipeline_mode": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "slice_counts": {
            "$ref": "#/components/schemas/SliceCountsDto"
          },
          "total_slices": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "project_id",
          "pipeline_mode",
          "slice_counts",
          "total_slices"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Mutagen workflow service. Spec is generated from handler annotations via utoipa per ADR-0002.",
    "license": {
      "name": ""
    },
    "title": "mutagen-service",
    "version": "0.3.3"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/projects/{project_id}/slices": {
      "get": {
        "operationId": "get_slices",
        "parameters": [
          {
            "description": "Project UUIDv7",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SliceQueueDto"
                }
              }
            },
            "description": "Slice queue snapshot"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown project_id"
          }
        },
        "tags": [
          "workflow"
        ]
      }
    },
...
    "/projects/{project_id}/state-log": {
      "get": {
        "operationId": "get_state_log",
        "parameters": [
          {
            "description": "Project UUIDv7",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Opaque pagination cursor (DSD-625)",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Soft cap on records; default 100, max 1000",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "nullable": true,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateLogPageDto"
                }
              }
            },
            "description": "State update log page"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown project_id"
          }
        },
        "tags": [
          "workflow"
        ]
      }
    },
    "/projects/{project_id}/status": {
      "get": {
        "operationId": "get_status",
        "parameters": [
          {
            "description": "Project UUIDv7",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusDto"
                }
              }
            },
            "description": "Workflow status snapshot"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown project_id"
          }
        },
        "tags": [
          "workflow"
        ]
      }
    },
...
  }
}